Fix bundle updates
It is possible for dependency cycles to be changed in both Gerrit
and GitHub without creating new patchsets.  In these cases, the queue
item bundles will not be updated.  This can lead to testing changes
with the wrong repo contents (particularly in the case where an item
is removed from a bundle).

Additionally, this caused errors loading the pipeline from ZK because
items may have had bundle objects associated with them which included
items no longer in the pipeline.  This particular case can only be
triggered when two items are removed from a pipeline ahead of the
remaining items in a bundle.  This is, in part, why the twe new
tests which have been added are fairly complex.

Change-Id: I04649a5b2b3143d8921f2390cda41895b1963687