Make lock cleanup more robust
The following sequence was observed:

1) Scheduler A locks a pipeline
2) Scheduler B attempts to lock the pipeline, getting as far
   as writing its lock contender file
3) ZK connections were suspended but not lost (sessions remained)
4) Scheduler A releases the lock
5) Scheduler B fails to delete its contender node due to the ZK
   connection loss.
6) All further attempts to lock the pipeline fail as they see
   scheduler B's lock contender node.

This happens because when deleting the contender node after an
unsuccessful lock attempt, the Lock recipe ignores all failures
and tries only once.

This change adjusts our session-aware lock mixin to use the usual
retry method to retry the operation if a retryable exception is
received.

[1] https://github.com/python-zk/kazoo/blob/92bd0c2a8540697633564deed683f87d1f611bf5/kazoo/recipe/lock.py#L322

Change-Id: I3ff4a4d6f47e8d35e133080e3e9a35c58ea503d2