Fix deduplication of child jobs in check
When we deduplicate jobs, we intend to call setResult on all of
the queue items with the deduplicated build.  This only worked
in dependent pipelines because we only looked for queue items in
the current bundle.  In independent pipelines, the queue items
can be in different bundles.

To resolve this, search for items with deduplicated builds in
across the whole queue in independent pipelines (using the approach
we use when deduplicating them to begin with).

Change-Id: I16436710c47b4f22df39e0cd82d0e289b2293c32