Fix deduplicating child jobs in check
If a second change is enqueued in check significantly after the
first, then node requests for child jobs may not be deduplicated
correctly.

Before deduplicating a build, Zuul applies parent data to child
jobs, then compares the child jobs to determine if they are
equivalent.  If so, then they are deduplicated.

This only happens one level in the hierarchy at a time.  Consider
the case where a change is enqueued and both the parent and child
jobs have completed (but the change is still in the queue waiting
on a third, unrelated, job).

If the second change in the bundle is enqueued, Zuul will:
1) Attempt to apply parent data to child jobs.
   Since no jobs have completed yet for this item, no parent data
   are applied.
2) Deduplicate jobs in the second change.
   Zuul will deduplicate the parent job at this point.
3) Zuul will compare the child jobs in the two changes and determine
   they are different because one has parent data and the other does
   not.
4) Zuul submits a node request for the child job.
5) On the next pipeline process, Zuul applies the parent data from
   the deduplicated parent job to the new child job.
6) Zuul deduplicates the child job, and the nodepool request is
   orphaned.

To correct this, we will repeat the process of applying parent data
to child jobs each time we find at least one build to deduplicate.

That means that all existing parent data will be applied to all jobs
on each pass through the pipeline processor no matter how deep the
dependency hierarchy.

Change-Id: Ifff17df40f0d59447f74cdde619246171279b553