Use a seeded PRNG when setting timer triggers
In the single scheduler days, we went to some effort to make sure
that we preserved apscheduler timer trigger jobs across reconfigurations
because to do otherwise risks missing zuul triggers if we reconfigure
within the jitter interval of a timer job.

In other words, if a timer trigger is set for midnight with a jitter
of 60s, and we reconfigure the tenant at midnight plus 10 seconds, if
the jitter would have caused the trigger to fire at midnight plus 30
seconds and the apscheduler job is canceled and recreated but this
time the jitter would cause it to fire at midnight plus 5 seconds,
the job would not fire.

Our mitigation for this was to preserve the apscheduler jobs across
reconfigurations, assuming nothing has changed.

In the multi-scheduler world, a similar problem arises when we perform
a rolling restart during a jitter interval.  Fixing this by recording
all of the jobs in ZK would be costly (as we create an apscheduler job
for each project-branch participating in a timer pipeline).  But we
can address the bulk of the problem just by ensuring that the same
jitter values are used on all schedulers.  This reduces the window for
missed jobs down to the time it takes for the schedulers to run the
timer trigger election (which, assuming an orderly shutdown, is very
fast).

In APScheduler the jitter value is calculated for every event.  This
new approach calculates a fixed jitter for a job, and will use that
value every time the job fires.  So if a job ends up running at
midnight plus 10 seconds one day, and no configuration changes,
it will run at midnight plus 10 seconds the next day.  This is
sufficient to achieve what we want to do with jitter: avoid starting
all the jobs at once.

We subclass the apscheduler CronTrigger to implement this custom
fixed jitter behavior and use a PRNG seeded with the tenant name so
that each scheduler will use the same value for each project-branch
combination, assuming they have the same configuration (and they
should).  We are implicitly relying on the dictionary ordering
in the layout, as well as branch list ordering to provide consistency
here.

Docs are updated to highlight that the same jitter value may be reused,
but no release note is added since the behavior change is not
significant.

Change-Id: Idd9d3a8cfa791860e46e4fc508566417f5d5a9bf