Add a bundle-changed safety check
Several recent bugs and attempted fixes have shown that there may
be some edge cases in the handling of dependency cycles that have
the potential to cause jobs to run with the wrong changes in place.

While we work on longer-term fixes to those, add a safety check to
the pipeline processor so that if we detect a change to the bundle
contents of a queue item, we remove the item from the queue.  We
may not necessarily perform the optimal behavior with this, but it
should keep us from running jobs with known incorrect changes.

This change requires some minor adjustment to some existing unit
tests (it doesn't significantly change the outcome, but it does
cause some jobs to be aborted sooner).  A followup change will add
some more tests which would fail without this change but merit
separate review.

Change-Id: Ia7b1d5b7e3d6910a709478082929f96364ca996b