Load config errors less often
If we encounter network delays loading the config errors endpoint
(which is likely if it's large), then we can end up issuing multiple
outstanding GET requests for the same endpoint.  Avoid that by using
the isFetching pattern used by other members of the store.

Also, store the tenant name for the most recent fetch so that we
*can* have multiple outstanding requests for different tenants, but
we only accept the most recent one.

Change-Id: Ic49027e137b8c30004d43059db259d10c5443812