Add tenant_status web endpoint
The config error list is getting longer, for everyone, especially
now that we are including warnings.  To avoid loading the entirety
when it's not necessary, add an API endpoint that simply returns
the number of config errors.  This can later be used by the web app
to highlight the blue bell without actually fetching the errors.

This endpoint is named tenant_status so that if we find more details
like this that we want to add in the future, we can extend the returned
dictionary with them.

It is not added to the tenant info endpoint because that is
unauthenticated and should not leak information about the tenant.

Change-Id: Ie11eb26dc38e28922ddabbca39d89cda7e763d13