Add zuul-scheduler status command
This outputs some basic status information to aid in debugging.

This isn't intended for use under normal circumstances.  But in
some cases, it is useful to know what the scheduler run loop is
doing without combing through logs or using SIGUSR2.  In these
cases, running "zuul-scheduler status" will now print out some
basic information about what two threads in the scheduler are
doing.

This requires some refactoring of the command handler to accept
the possibility of sending output back to the client.

Change-Id: Ia253d12a818174fbd26dccaf3e029284cb203a31