dnm: Demonstrate stable buildset UUID concept
The buildset defines the current position of an item in the queue. If
the item is moved or the sequence of items ahead changes, the buildset
is no longer valid.

This change proposes a new stable buildset UUID that can be used to
detect if a buildset needs to be reset, which is necessary when the item
ahead changed.

The stable UUID is calculated based on the stable UUID of the item ahead
in the queue or the last merged change for the item at the head of the
queue.

Change-Id: I6bf811ab4c1a959ef38f978fe070c1420df071a0