Update patch set 1
Patch Set 1: Code-Review-1

Can we add a test assertion for this?  I know we have a test for it already.

Bonus: we need a warning that we're not going to remove in order to test the warnings system.  Right now, most of those tests are using pipeline queue deprecation.  We could change them to use this.  (That's not a -1 for this change, that's just something we should do next.)

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1, 51dcffb7436f0e74ac18df3ca25a4d67944f39c1