Identify frozen job data by UUID instead of name
Change the frozen job data storage in ZK from being identified by name
to UUID. This allows us to handle multiple frozen jobs in a buildset
with the same name.

Since the job data path is opaque to the scheduler we don't need to add
any backward compatibility code.

This is done in preparation of the circular dependency refactoring as
detailed in I8c754689ef73ae20fd97ac34ffc75c983d4797b0.

Change-Id: Ic4df16e8e1ec6908234ecdf91fe08408182d05bb