wip: Don't schedule merge for non-change items
Currently we schedule a merge for every item that is added to a
pipeline. This happens before we check if there are any jobs to be
executed. This overhead can become quite substantial for projects with a
lot of branches that also have a periodic pipeline config, but only for
a small subset of those branches.

For branch/ref triggered items there doesn't seem to be a need to
schedule an initial merge as there is nothing to be merged and the repo
state will be included as part of the extra repo state later on.

Change-Id: I157ed52dba8f4e197b35798217b23ec7f035b2d9