Don't trigger timer events for dynamic branches
The documentation for the `always-dynamic-branches` regex [0] mentions
that no perodic jobs will run for matching branches. However, since the
timer driver did not get the list of project branches from the tenant
project config but directly from the source, we did not correctly
exclude those branches.

To fix this we'll use the branches from the tenant project config which
doesn't include always dynamic branches by default.

[0] https://zuul-ci.org/docs/zuul/latest/tenants.html#attr-tenant.untrusted-projects.%3Cproject%3E.always-dynamic-branches

Change-Id: I39f6a1a90504c11155dca70e3e9404b08b9de8ab