wip: Send job parent data + artifacts via build request
With job parents that supply data we might end up updating the (secret)
parent data and artifacts of a job multiple times in addition to also
storing duplicate data as most of this information is part of the
parent's build result.

Instead we will collect the parent data and artifacts before scheduling
a build request and send it as part of the request paramters.

TODO:
- add a model API upgrade test
- add backward compatibility

Change-Id: Ic4a85a57983d38f033cf63947a3b276c1ecc70dc