Add image debug build option
This adds an option in the Dockerfile to create a debug image.

It's not clear if the Zuul project itself should publish debug
images; if we do want to, it will be easy enough to do so with
this change.

Regardless, this will let users easily create their own debug
images when necessary.

Change-Id: I8aa7da5f77bb7b34c06b7aaaf2a7595a7926b67e