Force Nodepool re-election on connection suspended
When the Zookeeper connection is suspended we might miss some events
during the time until the client is able to reconnect. This can lead to
jobs waiting for node requests that are already fulfilled.

To fix this edge-case we force a re-election of the Nodepool event
watcher when the connection is suspended. This fixes the issue with lost
events as the event watcher will re-send nodes-provisioned events for
all ready requests when the election was won.

Change-Id: I69b39bb02481241d584253906922ae74b94060cf