Conditionally add index.html to the "view log" link
If the site has configured zuul manifest generation with
"index_links" set to true, they have indicated that the zuul web
ui needs to append "index.html" to all of the directory entries
listed in the zuul manifest when generating the "Logs" tab.

The same is almost certainly true for the raw link to the log
location (appearing as "View log" in the summary area), so once
we retrieve the manifest, if index_links is set there, then
set the "View log" link to also include an index.html.

If the manifest is missing, then we will fall back to returning
just the raw URL (and seeing the actual URL returned to Zuul may
be preferable when debugging log failures anyway). For most users,
most of the time, this should produce a reasonable "View log" link.

Change-Id: Ia63943742a7a20bbf271a4de692dd41f2fe0016d