Add index.html to the "view log" link
If the build log_url ends with a /, the best case is that the
log hosting platform has a default document name defined and will
serve index.html automatically (whether that is a real index.html
generated by the zuul log upload roles, or an autoindex).

Increasingly, however, it's the case that many platforms don't
support that (GCP and AWS both do not in some or all configurations).

To address this and make the "view logs" link correct more often
than it's wrong, let's append "index.html" if it ends with a /.

Change-Id: Ia63943742a7a20bbf271a4de692dd41f2fe0016d