Restore lingering deduplicated builds
We intend to handle the case where two queue items in check share
a deduplicated build and one of them completes and is dequeued while
the other continues.  To handle this, we avoid removing any queue
items (and therefore their associated builds) from ZK if their builds
show up in any currently-enqueued queue item.  However, we don't
actually have a mechanism to load a build from ZK in that situation
if it isn't in a queue item that is currently enqueued.

This change adds such a mechanism.

Now the program flow will be:

* Two changes enqueued in check
* At least one job/build in both deduplicated
* First change completes all jobs and is dequeued
* The pipeline cleanup method will leave the queue item in ZK even
  though the reference to it in the change queue is removed
* Second change continues running jobs
* When a scheduler loads the pipeline state, it will discover that
  change 2 references a build from change 1 that it has not deserialized,
  so it will explicitly deserialize that build
* Once change 2 completes, the normal pipeline cleanup method will
  delete both items.

This also updates the Build __repr__ method since some error reporting
uses that and it always assumes there is a build.job, however, if an
error happens early in deserialization, that attribute may not be set.

Change-Id: I86c159c82b858e67433bdaa1e479471b21ea8b86