Fix flaky deduplication test
The test test_job_deduplication_child_of_diff_parent has been flaky
because the following sequence can happen (some steps omitted):

* Two changes enqueued with a deduplicated job
* Pipeline processor runs and issues node request for deduplicated build
* Pipeline processor runs again and processes events related to
  first change
* Node request completes (after processing started processing first change)
* Pipeline processor continues and processes second change
* Pipeline processor attempts to start job for deduplicated build
  (an exception is raised)

The processor should ignore the completed node request on the second change
and allow the next pass to handle it when processing the first change.

To ensure this, we set the in-memory-only "ready to run" flag on the job
to False for any deduplicated job.  This way we will only start builds
when processing their actual queue item.

Because of the sequencing of this race condition (needing to return a build
completed event between processing two queue items in the same pipeline run),
it is impractical to add a test.

Change-Id: I8aa2a8adab273304bf518ca3def429dda7f8c544