Fix some API endpoints with special chars
None of our API endpoints expected URL quoting except one: the
/job/{jobname} endpoint.  This was previously updated to expect
URL quoting and the web UI was updated to send it.  This was done
mostly to accomodate job names with slashes.  Every other endpoint,
including /project/{projectname} which is notable for very commonly
including slashes, expects unquoted input.

Let's normalize this by making the /job/ endpoint just like the
others.  To do this, we update the routes configuration for both
the JS web app and zuul-web to expect slashes in the job name.

Additionally, the freeze-job API endpoint did not accept quoted
input (nor did it expect slashes) for the job name (but it did for
every other component) so adjust it to expect slashes as well.

The web JS did not expect slashes in semaphore names, so update it
as well.

The test is adjusted to include more special characters in more
places (including covering the freeze-job api).

A reasonable alternative would be to go the other way and support
url quoting everywhere.  That would require a much larger change
and behavior coordination between the JS routes component, the JS
Axios HTTP library, and the python routes dispatcher.  If we do
that, it should be considered as its own effort apart from this bugfix.

Change-Id: I3fb575b037823263b4def812168da8f343f7453b