Add safety when setting refs
The recent update to make setting refs more efficient could encounter
an edge case if a branch or tag was removed from the upstream repo
after the repo state was retrieved by zuul.  If removing the ref caused
the underlying objects to be removed not not be sent in a fetch, then
our blindly setting the repo state with a ref pointed to an unresolvable
object could leave the repository in a corrupted state.

To recover from any potential corruption that may have somehow happened,
this change adds an additional case where we will remove the underlying
repo and re-clone.

To prevent any such corruption from happening in the first place, we add
a check that each hexsha is resolvable before we set it when restoring
the repo state.  This does add a small amount of overhead, but should be
much less than manipulating the loos refs one at a time.  A copy of nova
with 10,000 refs adds 100ms for this checking.

Change-Id: Ifd298905e634f83a147644d35ff3ea1c143b3d1f