Implement server-side filtering and pagination of config errors
In order to support pagination of config errors (so that when a user
does decide to look at the config errors page, we don't necessarily
need to transfer all of the data each time, implement server-side
filtering and pagination.

A later change can implement the same in the web ui.

Change-Id: I0c6cb8a10cd4d807ed92cad438ef592b1cdaf19b