Fix some old entries in zuul_ref table
The recent db migration used existing build table entries to
populate the new zuul_ref table which is normalized so that it has
a single row for each (project, ref, change, patchset, oldrev, newrev)
combination.  Note that 'branch' is not included in that list because
a given ref implies a given branch.

However, some old (we're talking many, many, many years old) entries
in the builds table for periodic jobs were not made with a value in
the branch column.  This can cause problems with current jobs since
a periodic job today will use the same ref table entry as one from
10 years ago.  Periodic jobs have no change, patchset, oldrev, and
newrev fields, so the only unique attributes are project and ref.

When we collapsed the build table entries down to a single zuul_ref
row for each combination, if we picked one for a branch-tip ref as
described above with no value in the branch field, then as we
associate new builds with that ref, we won't be able to search for
them by branch (or even display the correct branch) since it will
be empty.

To correct this, update the zuul_ref table so that every row where
the branch is empty but the ref starts with 'refs/heads/' is updated
to include the branch (as read from the ref).

This migration can happen on-line with no coordination, and it should
take only a few seconds.

Change-Id: Idaa1ce91e9b763fd64e50f55bba4eb948ab21171