Further fixes to check deduplication
This fixes a similar case as I86c159c82b858e67433bdaa1e479471b21ea8b86
but at a different point in the pipeline lifecycle.

In both cases, a queue item in check has completed and been
dequeued, while a second item that deduplicated one of its jobs
remains.

We store both the node request and nodeset information on the buildset
object.  When builds are finished, we clear the node request from the
buildset object.  Depending on timing, we may have information about the
nodeset while we may or may not have the request id.

We use the presence of nodeset information to decide whether to run a
job, but we pass the node request to the executor for it to lock the
nodeset.

This patch handles the case where we have deduplicated nodeset information
but have already cleared the node request id.  In that case, we would
attempt to run the job a second time while passing in None as the
node request id.

To correct this, if we learn that we are unable to restore information
about a deduplicated build, clear all of the nodeset and node request
information about that job from the secondary buildset so that when it
attempts to re-run the build, it starts from a clean slate.

This case may be a superset of the other case which means we won't hit
the failsafe code added in I86c159c82b858e67433bdaa1e479471b21ea8b86.
That should be fine, and we will remove code related to both cases once
the circular dependency refactor is complete.

Change-Id: I95511763236d5a99c14477989d573bd5c1726b76