Set remote head refs in workspace repo clone
With the optimization in Ie8846c48ccd6255953f46640f5559bb41491d425 we no
longer keep the remote refs around. The remote origin ref in a workspace
repo can be used to find the diff of the changes being tested.

This behavior still works correctly for projects with a proposed change,
as we explicitly set the remote ref before performing any merges.
However, other projects that are part of the checkout but are not
changed in an item will no longer have any remote refs.

To restore the previous behavior we will also create the matching origin
ref for all processed head refs for workspace repos.

Change-Id: Ibf782412c49d44f37f32f234028d8203cf3e81b4